refactor: tidy up `cache-inline`
